+++ lib/socky-client.rb
@@ -7,12 +7,12 @@ require File.dirname(__FILE__) + '/socky-client/websocket'
 module Socky
 
   class << self
-    
+
     attr_accessor :config_path, :logger
     def config_path
       @config_path ||= 'socky_hosts.yml'
     end
-    
+
     def config
       @config ||= YAML.load_file(config_path).freeze
     end
@@ -29,11 +29,11 @@ module Socky
     def hosts
       config[:hosts]
     end
-    
+
     def logger
       @logger ||= Logger.new(STDOUT)
     end
-    
+
     def deprecation_warning(msg)
       logger.warn "DEPRECATION WARNING: " + msg.to_s
     end
@@ -55,48 +55,20 @@ module Socky
     end
 
     def send_message(data, opts = {})
-      to = opts[:to] || {}
-      except = opts[:except] || {}
-      
-      # Move to new syntax
-      if opts[:to] || opts[:except]
-        deprecation_warning "Using of :to and :except will be removed in next version - please move to new syntax."        
-      end
-      to[:client]   ||= opts[:client]
-      to[:clients]  ||= opts[:clients]
-      to[:channel]  ||= opts[:channel]
-      to[:channels] ||= opts[:channels]
-      # end of new syntax
-
-      unless to.is_a?(Hash) && except.is_a?(Hash)
-        raise "recipiend data should be in hash format"
-      end
-
-      to_clients = to[:client] || to[:clients]
-      to_channels = to[:channel] || to[:channels]
-      except_clients = except[:client] || except[:clients]
-      except_channels = except[:channel] || except[:channels]
+      clients = opts[:client] || opts[:clients]
+      channels = opts[:channel] || opts[:channels]
 
       # If clients or channels are non-nil but empty then there's no users to target message
-      return if (to_clients.is_a?(Array) && to_clients.empty?) || (to_channels.is_a?(Array) && to_channels.empty?)
+      return if (clients.is_a?(Array) && clients.empty?) || (channels.is_a?(Array) && channels.empty?)
 
       hash = {
         :command  => :broadcast,
-        :body     => data,
-        :to => {
-          :clients  => to_clients,
-          :channels => to_channels,
-        },
-        :except => {
-          :clients  => except_clients,
-          :channels => except_channels,
-        }
+        :data     => data,
+        :clients  => clients,
+        :channels => channels
       }
 
-      [:to, :except].each do |type|
-        hash[type].reject! { |key,val| val.nil? || (type == :except && val.empty?)}
-        hash.delete(type) if hash[type].empty?
-      end
+      hash.reject! { |key,val| val.nil? }
 
       send_data(hash)
     end
@@ -104,7 +76,7 @@ module Socky
     def send_query(type)
       hash = {
         :command  => :query,
-        :type     => type
+        :data     => type
       }
       send_data(hash, true)
     end
@@ -124,7 +96,7 @@ module Socky
         end
       end
       if response
-        res.collect {|r| JSON.parse(r)["body"] }
+        res.collect {|r| JSON.parse(r)["data"] }
       else
         true
       end